requirements.txt
cartooner
rembg
Pillow
streamlit
numba
numpy
opencv-python-headless
onnxruntime